split tests into separate module dir
