+++ shielder/contract/drink_tests/utils/mod.rs
@@ -1,0 +1,2 @@
+pub mod psp22;
+pub mod shielder;
